tests/server/test-requirements.txt
pytest
cattle==0.4.0
flake8
docker-py
requests==2.6.0
